+++ frontend/src/web/log/LogFileAnalysis.js
@@ -159,21 +159,33 @@ class LogFileAnalysisCore extends React.PureComponent {
         {!incident &&
           <Container fullHeight className="overflow-y-auto">
             <Tile isParent isFluid style={{ paddingLeft: 0, paddingRight: 0 }}>
-              {incidentList.map(ic => (
-                <Tile
-                  key={`${projectId}-${ic.id}`} className="incident-tile"
-                  onClick={this.handleIncidentClick(ic.id)}
-                >
-                  <Box isLink>
-                    <div className="content">
-                      <div className="label">Start Time</div>
-                      <div>{moment(ic.incidentStartTime).format('YYYY/MM/DD mm:ss')}</div>
-                      <div className="label">End Time</div>
-                      <div>{moment(ic.incidentEndTime).format('YYYY/MM/DD mm:ss')}</div>
+            {incidentList.map(ic => (
+              <Tile
+                key={`${projectId}-${ic.id}`} className="incident-tile"
+                onClick={this.handleIncidentClick(ic.id)}
+              >
+                <Box isLink>
+                  <div className="content">
+                    <div>
+                      <div className="label" style={{ display: 'inline-block' }}>Start Time:</div>
+                      <div style={{ float: 'right' }}>{moment(ic.incidentStartTime).format('YYYY/MM/DD hh:mm')}</div>
                     </div>
-                  </Box>
-                </Tile>
-              ))}
+                    <div>
+                      <div className="label" style={{ display: 'inline-block' }}>End Time:</div>
+                      <div style={{ float: 'right' }}>{moment(ic.incidentEndTime).format('YYYY/MM/DD hh:mm')}</div>
+                    </div>
+                    <div>
+                      <div className="label" style={{ display: 'inline-block' }}>Clusters:</div>
+                      <div style={{ float: 'right' }}></div>
+                    </div>
+                    <div>
+                      <div className="label" style={{ display: 'inline-block' }}>Rare Events:</div>
+                      <div style={{ float: 'right' }}></div>
+                    </div>
+                  </div>
+                </Box>
+              </Tile>
+            ))}
             </Tile>
           </Container>
         }